ik: more porteer stuff to its own module
